fix validator

+++ internal/domain/accounts/get_account_balance_test.go
@@ -13,13 +13,13 @@ type mockGetAccountBalanceRepository struct {
 	err     error
 }
 
-func (m *mockGetAccountBalanceRepository) FindByID(ctx context.Context, accountID string) (float64, error) {
+func (m *mockGetAccountBalanceRepository) FindByID(ctx context.Context, accountID string) (int64, error) {
 	return m.account.Balance, m.err
 }
 
 func TestGetAccountBalance(t *testing.T) {
 	mockRepo := &mockGetAccountBalanceRepository{
-		account: &entities.Account{ID: "1", Name: "John Doe", CPF: "12345678900", Balance: 100.0},
+		account: &entities.Account{ID: "1", Name: "John Doe", CPF: "12345678900", Balance: 100},
 		err:     nil,
 	}
 
@@ -27,5 +27,5 @@ func TestGetAccountBalance(t *testing.T) {
 	output, err := uc.GetAccountBalance(context.Background(), "1")
 
 	assert.NoError(t, err)
-	assert.Equal(t, output.Balance, 100.0)
+	assert.Equal(t, output.Balance, 100)
 }
